Scroll Function
Added a dropdown menu for category when adding items manually to either shopping list or inventory. It has scroll and search functionality. I did not add a function to allow users to type in a category because the categories there should be enough for all possible grocery items but lmk if i should add that.