Updated Some Things in Shopping List and Inventory
Added a drop down menu for category in both shopping list table and inventory.
When an item is scanned, the scan code will try to match the open food facts categories of that item to one of the categories in the drop down menu. If there is not match then its automatically placed as other.